backend/file_sharing_app/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
Django==4.0.5
django-celery-beat==2.3.0
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hiredis==2.0.0
httpagentparser==1.9.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
openapi-codec==1.3.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
